M2S9/aula2707/pasta1/src/app/config/__init__.py
import os
from dotenv import load_dotenv, find_dotenv
load_dotenv(find_dotenv())


class Development(object):
    DEBUG = True
    TESTING = False
class Production(object):
    DEBUG = False
    TESTING = False
app_config = { 'development': Development, 'production': Production}
