M2S9/aula2707/pasta1/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Maicon Silva <smaicon@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
